Refactor amm_contracts structure
