+++ tests/amm_contracts/mod.rs
@@ -1,0 +1,3 @@
+pub mod amm;
+pub mod amm_proxy;
+pub mod utils;
