+++ tests/amm_contracts/utils.rs
